+++ src/main/java/roomescape/ReservationDto.java
@@ -2,6 +2,7 @@ package roomescape;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
+import roomescape.domain.Reservation;
 
 public record ReservationDto(
         Long id,
