+++ src/main/java/roomescape/controller/ReservationController.java
@@ -8,9 +8,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.Reservation;
+import roomescape.domain.Reservation;
 import roomescape.ReservationDto;
-import roomescape.ReservationRepository;
+import roomescape.domain.ReservationRepository;
 
 @RestController
 public class ReservationController {
